Caso de Prueba 2, los lados decimales son permitidos

+++ main.py
@@ -9,6 +9,7 @@ def clasificar_triangulo(lado1, lado2, lado3):
     Un string que indica si el triángulo es escaleno,
     isósceles, equilátero o inválido.
     """
+
     #Se verifica que los valores ingresados no sea o sean 0 ni negativos
     if(lado1 <= 0 or lado2 <= 0 or lado3 <=0):
         if(lado1 == 0 or lado2 == 0 or lado3 ==0):
@@ -27,9 +28,9 @@ def clasificar_triangulo(lado1, lado2, lado3):
         return "Triángulo escaleno"
 
 # Solicitar al usuario la longitud de los lados
-lado1 = int(input("Ingrese la longitud del primer lado: "))
-lado2 = int(input("Ingrese la longitud del segundo lado: "))
-lado3 = int(input("Ingrese la longitud del tercer lado: "))
+lado1 = float(input("Ingrese la longitud del primer lado: "))
+lado2 = float(input("Ingrese la longitud del segundo lado: "))
+lado3 = float(input("Ingrese la longitud del tercer lado: "))
 
 # Clasificar el triángulo y mostrar el resultado
 clasificacion = clasificar_triangulo(lado1, lado2, lado3)
